+++ plugins/woocommerce-admin/packages/components/src/select-control/list.js
@@ -100,7 +100,9 @@ class List extends Component {
 				break;
 
 			case ENTER:
-				this.select( options[ selectedIndex ] );
+				if ( options[ selectedIndex ] ) {
+					this.select( options[ selectedIndex ] );
+				}
 				event.preventDefault();
 				event.stopPropagation();
 				break;
@@ -116,9 +118,10 @@ class List extends Component {
 				return;
 
 			case TAB:
-				this.select( options[ selectedIndex ] );
+				if ( options[ selectedIndex ] ) {
+					this.select( options[ selectedIndex ] );
+				}
 				setExpanded( false );
-				onSearch( null );
 				break;
 
 			default:
@@ -139,6 +142,10 @@ class List extends Component {
 	}
 
 	componentDidMount() {
+		const { selectedIndex } = this.props;
+		if ( selectedIndex > -1 ) {
+			this.scrollToOption( selectedIndex );
+		}
 		this.toggleKeyEvents( true );
 	}