+++ plugins/woocommerce-admin/packages/components/src/select-control/test/index.js
@@ -201,6 +201,25 @@ describe( 'SelectControl', () => {
 		expect( getAllByRole( 'option' ) ).toHaveLength( 1 );
 	} );
 
+	it( 'should not automatically select first option on focus', async () => {
+		const onChangeMock = jest.fn();
+		const { getByRole, queryByRole } = render(
+			<SelectControl
+				isSearchable
+				showAllOnFocus
+				options={ options }
+				onSearch={ () => options }
+				onFilter={ () => options }
+				onChange={ onChangeMock }
+			/>
+		);
+		getByRole( 'combobox' ).focus();
+		await waitFor( () =>
+			expect( getByRole( 'option', { name: 'bar' } ) ).toBeInTheDocument()
+		);
+		expect( queryByRole( 'option', { selected: true } ) ).toBeNull();
+	} );
+
 	describe( 'selected value', () => {
 		it( 'should return an array if array', async () => {
 			const onChangeMock = jest.fn();
@@ -258,6 +277,72 @@ describe( 'SelectControl', () => {
 				options[ 2 ].key,
 				'test'
 			);
+		} );
+
+		describe( 'prop excludeSelectedOptions', () => {
+			it( 'should preserve selected option when focused', async () => {
+				const onChangeMock = jest.fn();
+				const { getByRole } = render(
+					<SelectControl
+						isSearchable
+						showAllOnFocus
+						selected={ options[ 2 ].key }
+						options={ options }
+						onSearch={ () => options }
+						onFilter={ () => options }
+						onChange={ onChangeMock }
+						excludeSelectedOptions={ false }
+					/>
+				);
+				getByRole( 'combobox' ).focus();
+				await waitFor( () =>
+					expect(
+						getByRole( 'option', { name: 'bar' } )
+					).toBeInTheDocument()
+				);
+				// In browser, the <Button> in <List> component is automatically "selected" when <Control> lost focus.
+				// I was not able to produce the same behaviour with unit test, but a click on the currently
+				// selected option should be sufficient to simulate the logic in this test.
+				userEvent.click( getByRole( 'option', { selected: true } ) );
+				expect( onChangeMock ).toHaveBeenCalledTimes( 0 );
+			} );
+
+			it( 'should reset selected option if searching', async () => {
+				const onChangeMock = jest.fn();
+				const { getByRole, queryByRole } = render(
+					<SelectControl
+						isSearchable
+						showAllOnFocus
+						selected={ options[ 2 ].key }
+						options={ options }
+						onSearch={ () => options }
+						onFilter={ () => options }
+						onChange={ onChangeMock }
+						excludeSelectedOptions={ false }
+					/>
+				);
+
+				getByRole( 'combobox' ).focus();
+				await waitFor( () =>
+					expect(
+						getByRole( 'option', { name: 'bar' } )
+					).toBeInTheDocument()
+				);
+
+				userEvent.click( getByRole( 'option', { name: 'bar' } ) );
+				userEvent.clear( getByRole( 'combobox' ) );
+				userEvent.type( getByRole( 'combobox' ), 'bar' );
+
+				await waitFor( () =>
+					expect(
+						getByRole( 'option', { name: 'bar' } )
+					).toBeInTheDocument()
+				);
+
+				expect(
+					queryByRole( 'option', { selected: true } )
+				).toBeNull();
+			} );
 		} );
 
 		it( 'disables the component', async () => {
@@ -323,5 +408,152 @@ describe( 'SelectControl', () => {
 		);
 
 		expect( getByText( options[ 1 ].label ) ).toBeInTheDocument();
+	} );
+
+	describe( 'keyboard interaction', () => {
+		it( 'pressing keydown on combobox should highlight first option', async () => {
+			const { getByRole } = render(
+				<SelectControl
+					options={ options }
+					onSearch={ () => options }
+					onFilter={ () => options }
+					selected={ null }
+				/>
+			);
+
+			getByRole( 'combobox' ).focus();
+			await waitFor( () =>
+				expect(
+					getByRole( 'option', { name: 'bar' } )
+				).toBeInTheDocument()
+			);
+
+			userEvent.type( getByRole( 'combobox' ), '{arrowdown}' );
+
+			expect(
+				getByRole( 'option', { selected: true } ).textContent
+			).toEqual( options[ 0 ].label );
+		} );
+
+		it( 'pressing keydown on combobox with selected should highlight the next option', async () => {
+			const { getByRole } = render(
+				<SelectControl
+					options={ options }
+					onSearch={ () => options }
+					onFilter={ () => options }
+					selected={ options[ 1 ].key }
+					excludeSelectedOptions={ false }
+				/>
+			);
+
+			getByRole( 'combobox' ).focus();
+			await waitFor( () =>
+				expect(
+					getByRole( 'option', { name: 'bar' } )
+				).toBeInTheDocument()
+			);
+
+			userEvent.type( getByRole( 'combobox' ), '{arrowdown}' );
+
+			expect(
+				getByRole( 'option', { selected: true } ).textContent
+			).toEqual( options[ 2 ].label );
+		} );
+
+		it( 'pressing keydown on combobox with selected last option should rotate to first option', async () => {
+			const { getByRole } = render(
+				<SelectControl
+					options={ options }
+					onSearch={ () => options }
+					onFilter={ () => options }
+					selected={ options[ options.length - 1 ].key }
+					excludeSelectedOptions={ false }
+				/>
+			);
+
+			getByRole( 'combobox' ).focus();
+			await waitFor( () =>
+				expect(
+					getByRole( 'option', { name: 'bar' } )
+				).toBeInTheDocument()
+			);
+
+			userEvent.type( getByRole( 'combobox' ), '{arrowdown}' );
+
+			expect(
+				getByRole( 'option', { selected: true } ).textContent
+			).toEqual( options[ 0 ].label );
+		} );
+
+		it( 'pressing keyup on combobox should highlight last option', async () => {
+			const { getByRole } = render(
+				<SelectControl
+					options={ options }
+					onSearch={ () => options }
+					onFilter={ () => options }
+					selected={ null }
+				/>
+			);
+
+			getByRole( 'combobox' ).focus();
+			await waitFor( () =>
+				expect(
+					getByRole( 'option', { name: 'bar' } )
+				).toBeInTheDocument()
+			);
+
+			userEvent.type( getByRole( 'combobox' ), '{arrowup}' );
+
+			expect(
+				getByRole( 'option', { selected: true } ).textContent
+			).toEqual( options[ options.length - 1 ].label );
+		} );
+
+		it( 'pressing tab on combobox should hide option list', async () => {
+			const { getByRole, queryByRole } = render(
+				<SelectControl
+					options={ options }
+					onSearch={ () => options }
+					onFilter={ () => options }
+					selected={ null }
+				/>
+			);
+
+			getByRole( 'combobox' ).focus();
+			await waitFor( () =>
+				expect(
+					getByRole( 'option', { name: 'bar' } )
+				).toBeInTheDocument()
+			);
+
+			userEvent.type( getByRole( 'combobox' ), '{tab}' );
+
+			expect( queryByRole( 'option', { selected: true } ) ).toBeNull();
+		} );
+
+		it( 'pressing enter should select highlighted option', async () => {
+			const onChangeMock = jest.fn();
+			const { getByRole } = render(
+				<SelectControl
+					options={ options }
+					onSearch={ () => options }
+					onFilter={ () => options }
+					selected={ null }
+					excludeSelectedOptions={ false }
+					onChange={ onChangeMock }
+				/>
+			);
+
+			getByRole( 'combobox' ).focus();
+			await waitFor( () =>
+				expect(
+					getByRole( 'option', { name: 'bar' } )
+				).toBeInTheDocument()
+			);
+
+			userEvent.type( getByRole( 'combobox' ), '{arrowdown}{enter}' );
+
+			expect( onChangeMock ).toHaveBeenCalled();
+		} );
 	} );
 } );